rename file to options
