+++ libraries/azure-app-configuration-importer/tests/featureFlagConfigurationSettingsConverter.spec.ts
@@ -11,7 +11,7 @@ import * as fs from "fs";
 import { ConfigurationFormat } from "../src/enums";
 import { ArgumentError } from "../src/errors";
 import { StringConfigurationSettingsSource } from "../src/settingsImport/stringConfigurationSettingsSource";
-import { StringSourceOptions } from "../src/importOptions";
+import { StringSourceOptions } from "../src/options";
 import { assertThrowAsync } from "./utlis";
 import { MsFeatureFlagValue } from "../src/featureFlag";
