+++ libraries/azure-app-configuration-importer/tests/readableStreamConfigurationSettingsSource.spec.ts
@@ -11,7 +11,7 @@ import { AppConfigurationImporter } from "../src/appConfigurationImporter";
 import { MockUpHttpHeaderLike, MockupResourceLike } from "./appConfigurationImporter.spec";
 import { ConfigurationFormat, ConfigurationProfile } from "../src/enums";
 import { ReadableStreamConfigurationSettingsSource } from "../src/settingsImport/readableStreamConfigurationSettingsSource";
-import  { ReadableStreamSourceOptions } from "../src/importOptions";
+import  { ReadableStreamSourceOptions } from "../src/options";
 import { assertThrowAsync, listConfigurationSettings } from "./utlis";
 import { ParseError } from "../src/errors";
